feat: data sending to websocket

+++ raspberry_grove_pi_service/schemas.py
@@ -2,7 +2,7 @@ from pydantic import BaseModel
 
 
 class SensorData(BaseModel):
-    temperature: int | None = None
-    humidity: int | None = None
+    temperature: float | None = None
+    humidity: float | None = None
     sound: int | None = None
     light: int | None = None
